test: colocate benchmarks with the code they exercise
Distribute the benchmarks from benchmark_test.go into the _test.go file of
the package code each one exercises, matching this repo's per-file test
layout (idiomatic Go; the standard library colocates benchmarks too):

  BenchmarkDecodeDomain                  -> decoder_test.go
  BenchmarkEscapePathClean/Dirty         -> request_test.go
  BenchmarkCleanStringClean/Dirty        -> print_test.go
  BenchmarkVCardValues                   -> vcard_test.go

BenchmarkDecodeDomain now loads its fixture via test.LoadFile for
consistency (outside the timed loop, so results are unaffected).
