+++ print_test.go
@@ -32,3 +32,37 @@ func loadObject(filename string) RDAPObject {
 
 	return result
 }
+
+// cleanString runs on every printed heading and value. The clean case (no bad
+// runes) is the common case and should avoid the rune-by-rune strings.Map scan.
+var (
+	benchCleanStringInputs = []string{
+		"Domain Name",
+		"EXAMPLE.COM",
+		"2021-01-01T00:00:00Z",
+		"client transfer prohibited",
+		"https://rdap.verisign.com/com/v1/domain/EXAMPLE.COM",
+		"ns1.example.com",
+		"registrant",
+		"Registrar Abuse Contact Email",
+	}
+	benchCleanStringDirty = "line one\nline two\r\x00trailing"
+)
+
+func BenchmarkCleanStringClean(b *testing.B) {
+	p := &Printer{}
+	b.ReportAllocs()
+	for range b.N {
+		for _, s := range benchCleanStringInputs {
+			_ = p.cleanString(s)
+		}
+	}
+}
+
+func BenchmarkCleanStringDirty(b *testing.B) {
+	p := &Printer{}
+	b.ReportAllocs()
+	for range b.N {
+		_ = p.cleanString(benchCleanStringDirty)
+	}
+}
