+++ request_test.go
@@ -244,3 +244,24 @@ func TestNewAutoRequest(t *testing.T) {
 		}
 	}
 }
+
+// escapePath runs on every request path. The clean case (no byte needs
+// escaping) is overwhelmingly common and should not allocate.
+var (
+	benchEscapePathClean = "rdap.example.com"
+	benchEscapePathDirty = "xn--n3h.example/path with spaces & symbols"
+)
+
+func BenchmarkEscapePathClean(b *testing.B) {
+	b.ReportAllocs()
+	for range b.N {
+		_ = escapePath(benchEscapePathClean)
+	}
+}
+
+func BenchmarkEscapePathDirty(b *testing.B) {
+	b.ReportAllocs()
+	for range b.N {
+		_ = escapePath(benchEscapePathDirty)
+	}
+}
